refactor(widgets): extract FactsStrip from DrinkHeroPanel
Pull the private facts-strip container and _FactCell out of
DrinkHeroPanel into a shared, public FactsStrip/FactCell pair in
facts_strip.dart. FactsStrip now owns the divider colour and the
top/bottom/left border logic, so callers just supply the cells.

This is a pure refactor: DrinkHeroPanel renders the same three cells
it built before, pixel-identical, and the drink-detail goldens are
unchanged. The extraction exists so the brewery and style heroes can
reuse the same facts grammar instead of each reinventing it.
